# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/cancel_late_logistics": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for canceling or marking late logistic record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logistic"
                ],
                "summary": "Cancel or late logistics",
                "parameters": [
                    {
                        "description": "Logistic data",
                        "name": "logistic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swag.CancelLogistic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/v1/logistics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for retrieving all logistic records with pagination and search",
                "tags": [
                    "logistic"
                ],
                "summary": "Get all logistic records",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of logistics per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Post",
                        "name": "post",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Driver Type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Driver Position",
                        "name": "position",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Driver Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Location",
                        "name": "location",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "state",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetAllLogisticsResp"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for creating a new logistic record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logistic"
                ],
                "summary": "Create a logistic record",
                "parameters": [
                    {
                        "description": "Logistic data",
                        "name": "logistic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swag.CreateUpdateLogistic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseId"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/logistics/{logistic_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for retrieving a logistic record by ID",
                "tags": [
                    "logistic"
                ],
                "summary": "Get a logistic record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Logistic ID",
                        "name": "logistic_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Logistic"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for updating a logistic record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logistic"
                ],
                "summary": "Update a logistic record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Logistic ID",
                        "name": "logistic_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Logistic data",
                        "name": "logistic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swag.CreateUpdateLogistic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for deleting a logistic record",
                "tags": [
                    "logistic"
                ],
                "summary": "Delete a logistic record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Logistic ID",
                        "name": "logistic_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/logistics_with_cargo/{logistic_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for updating a logistic record with a cargo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logistic"
                ],
                "summary": "Update a logistic record with a cargo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Logistic ID",
                        "name": "logistic_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Logistic data",
                        "name": "logistic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swag.UpdateLogisticWithCargo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseId"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.GetAllLogisticsResp": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "logistics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LogisticResponse"
                    }
                }
            }
        },
...
        "models.ResponseError": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                }
            }
        },
        "models.ResponseId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "models.ResponseSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "swag.CancelLogistic": {
            "type": "object",
            "properties": {
                "cancel": {
                    "type": "boolean"
                },
                "company": {
                    "type": "string"
                },
                "logistic_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "section": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "whose_fault": {
                    "type": "string"
                }
            }
        },
...
        "swag.CreateUpdateLogistic": {
            "type": "object",
            "properties": {
                "cargo_id": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "notion": {
                    "type": "string"
                },
                "post": {
                    "type": "boolean"
                },
                "st_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}